+++ test/unleash.test.js
@@ -24,7 +24,7 @@ const defaultToggles = [
 ];
 function mockNetwork(toggles = defaultToggles, url = getUrl()) {
     nock(url)
-        .get('/features')
+        .get('/client/features')
         .reply(200, { features: toggles });
     return url;
 }
@@ -122,7 +122,7 @@ test('should re-emit events from repository and metrics', t => {
 test.cb('repository should surface error when invalid basePath', t => {
     const url = 'http://unleash-surface.app/';
     nock(url)
-        .get('/features')
+        .get('/client/features')
         .delay(100)
         .reply(200, { features: [] });
     const backupPath = join(tmpdir(), `test-tmp-${Math.round(Math.random() * 100000)}`);